requirements.txt
Django==1.9.6
wheel==0.24.0
PyMySQL==0.7.1
https://github.com/a1tus/sphinxapi-py3/zipball/master
django-ckeditor==5.0.3